Casting unsigned int to int

+++ more_malloc_free/1-string_nconcat.c
@@ -22,7 +22,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	}
 	
 	firstLength = strlen(s1);
-	concatPointer = (char *)malloc(((int)firstLength + n + 1) * sizeof(char));
+	concatPointer = (char *)malloc(((int)firstLength + (int)n + 1) * sizeof(char));
 
 	if (concatPointer == NULL)
 	{
@@ -36,7 +36,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	}
 
 	secondaryIndex = 0;
-	for (index = firstLength; index < ((int)firstLength + n); index++)
+	for (index = firstLength; index < ((int)firstLength + (int)n); index++)
 	{
 		concatPointer[index] = s2[secondaryIndex];
 		secondaryIndex++;
